tatystore-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.6
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.9
apscheduler==3.10.4
aiofiles==23.2.1
python-dotenv==1.0.0
email-validator
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
PyJWT==2.8.0
slowapi==0.1.9
python-json-logger>=2.0.7
